# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Coziyoo API v1
  version: 1.0.0
  description: Versioned API contract for Coziyoo marketplace backend.
servers:
  - url: http://localhost:3000
    description: Local
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /v1/auth/login:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: Logged in }
        '401': { description: Invalid credentials }
  /v1/auth/refresh:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: Refreshed }
...
  /v1/admin/auth/login:
    post:
      tags: [admin-auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: Admin logged in }
  /v1/admin/auth/refresh:
    post:
      tags: [admin-auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200': { description: Refreshed }
...
  /v1/admin/metadata/tables/{tableKey}/fields:
    get:
      tags: [admin-metadata]
      parameters:
        - in: path
          name: tableKey
          required: true
          schema: { type: string }
      responses:
        '200': { description: Dynamic field metadata for admin grids }
  /v1/admin/table-preferences/{tableKey}:
    get:
      tags: [admin-metadata]
      parameters:
        - in: path
          name: tableKey
          required: true
          schema: { type: string }
      responses:
        '200': { description: Admin table preferences }
    put:
      tags: [admin-metadata]
      parameters:
        - in: path
          name: tableKey
          required: true
          schema: { type: string }
      responses:
        '200': { description: Admin table preferences updated }